requirements.txt
# Reproducible baseline (tested in this repo)
# Recommended Python: 3.9-3.10
jax==0.4.8
jaxlib==0.4.7
optax==0.1.4
numpy==1.26.4
matplotlib==3.10.0
tqdm==4.67.1
PyYAML==6.0.2
